Move sed-builder Ws.java and App.java test files to iris-common. Refactor sed-builder tests.
