+++ sed-builder/src/test/java/cfa/vo/sed/builder/SegmentImporterTest.java
@@ -23,10 +23,10 @@ package cfa.vo.sed.builder;
 
 import cfa.vo.sed.setup.SetupManager;
 import cfa.vo.sed.setup.ISetup;
-import cfa.vo.sed.test.App;
+import cfa.vo.iris.test.App;
 import cfa.vo.sed.test.ConfigFactory;
 import cfa.vo.sed.test.URLTestConverter;
-import cfa.vo.sed.test.Ws;
+import cfa.vo.iris.test.Ws;
 import cfa.vo.sedlib.Sed;
 import cfa.vo.sedlib.Segment;
 import cfa.vo.sedlib.io.SedFormat;
