+++ sed-builder/src/test/java/cfa/vo/sed/filters/FITSFilterTest.java
@@ -23,9 +23,9 @@ import cfa.vo.sed.setup.SetupBean;
 import cfa.vo.sed.setup.validation.AxesValidator;
 import cfa.vo.sed.setup.validation.ErrorValidator;
 import cfa.vo.sed.setup.validation.IValidator;
-import cfa.vo.sed.test.App;
+import cfa.vo.iris.test.App;
 import cfa.vo.sed.test.URLTestConverter;
-import cfa.vo.sed.test.Ws;
+import cfa.vo.iris.test.Ws;
 import cfa.vo.sedlib.Segment;
 import org.junit.After;
 import org.junit.AfterClass;
