+++ sed-builder/src/test/java/cfa/vo/sed/filters/PhotometryCatalogTest.java
@@ -34,9 +34,9 @@ import cfa.vo.iris.sed.quantities.SPVYQuantity;
 import cfa.vo.iris.sed.quantities.YUnit;
 import cfa.vo.sed.setup.PhotometryCatalogBuilder;
 import cfa.vo.sed.setup.PhotometryPointBuilder;
-import cfa.vo.sed.test.App;
+import cfa.vo.iris.test.App;
 import cfa.vo.iris.test.Oracle;
-import cfa.vo.sed.test.Ws;
+import cfa.vo.iris.test.Ws;
 import cfa.vo.sedlib.io.SedFormat;
 
 import java.util.logging.Logger;
