+++ sed-builder/src/test/java/cfa/vo/sed/science/stacker/SedStackTest.java
@@ -26,8 +26,8 @@ import cfa.vo.sed.builder.AsciiConf;
 import cfa.vo.sed.builder.SedBuilder;
 import cfa.vo.sed.builder.SegmentImporter;
 import cfa.vo.sed.setup.SetupBean;
-import cfa.vo.sed.test.App;
-import cfa.vo.sed.test.Ws;
+import cfa.vo.iris.test.App;
+import cfa.vo.iris.test.Ws;
 import cfa.vo.sedlib.Segment;
 import java.net.URL;
 import java.util.ArrayList;
