--- sed-builder/src/test/java/cfa/vo/sed/test/Ws.java
@@ -1,81 +1,0 @@
-/**
- * Copyright (C) 2012 Smithsonian Astrophysical Observatory
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
-package cfa.vo.sed.test;
-
-import cfa.vo.iris.IWorkspace;
-import cfa.vo.iris.sed.ISedManager;
-import cfa.vo.iris.sed.SedlibSedManager;
-import cfa.vo.iris.units.UnitsManager;
-import cfa.vo.utils.Default;
-
-import java.util.ArrayList;
-import java.util.List;
-import javax.swing.JDesktopPane;
-import javax.swing.JFileChooser;
-import javax.swing.JFrame;
-import javax.swing.JInternalFrame;
-
-/**
- *
- * @author olaurino
- */
-public class Ws implements IWorkspace {
-
-    SedlibSedManager manager = new SedlibSedManager();
-
-//    JDesktopPane desktop = new JDesktopPane();
-    List<JInternalFrame> frames = new ArrayList<JInternalFrame>();
-
-    @Override
-    public void addFrame(JInternalFrame frame) {
-        frames.add(frame);
-    }
-
-    @Override
-    public void removeFrame(JInternalFrame frame) {
-        frames.remove(frame);
-    }
-
-    @Override
-    public JFileChooser getFileChooser() {
-        return null;
-    }
-
-    @Override
-    public ISedManager getSedManager() {
-        return manager;
-    }
-
-    @Override
-    public JFrame getRootFrame() {
-        return null;
-    }
-
-    @Override
-    public JDesktopPane getDesktop() {
-        return null;
-    }
-
-    @Override
-    public UnitsManager getUnitsManager() {
-        return Default.getInstance().getUnitsManager();
-    }
-}
